requirements.txt
Flask
flask-cors
python-dotenv
langchain
langchain-community
langchain-together
langchain-huggingface
tiktoken
chromadb
pypdf